remove PredicateFinder
