--- shardingsphere-sql-parser/shardingsphere-sql-parser-relation/src/test/java/org/apache/shardingsphere/sql/parser/relation/predicate/PredicateFinderTest.java
@@ -1,100 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.sql.parser.relation.predicate;
-
-import com.google.common.base.Optional;
-import org.apache.shardingsphere.sql.parser.relation.metadata.RelationMetas;
-import org.apache.shardingsphere.sql.parser.sql.segment.dml.column.ColumnSegment;
-import org.apache.shardingsphere.sql.parser.sql.segment.dml.predicate.PredicateSegment;
-import org.apache.shardingsphere.sql.parser.sql.segment.dml.predicate.value.PredicateRightValue;
-import org.apache.shardingsphere.sql.parser.sql.segment.generic.AliasSegment;
-import org.apache.shardingsphere.sql.parser.sql.segment.generic.OwnerSegment;
-import org.apache.shardingsphere.sql.parser.sql.segment.generic.TableSegment;
-import org.apache.shardingsphere.sql.parser.sql.value.identifier.IdentifierValue;
-import org.junit.Test;
-
-import java.util.Arrays;
-import java.util.Collections;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-public final class PredicateFinderTest {
-    
-    @Test
-    public void assertFindTableNameWhenSingleTable() {
-        TableSegment tableSegment = createTableSegment("table_1", "tbl_1");
-        PredicateSegment predicateSegment = createPredicateSegment(createColumnSegment());
-        Optional<String> actual = new PredicateFinder(mock(RelationMetas.class), Collections.singletonList(tableSegment)).findTableName(predicateSegment);
-        assertTrue(actual.isPresent());
-        assertThat(actual.get(), is("table_1"));
-    }
-    
-    @Test
-    public void assertFindTableNameWhenColumnSegmentOwnerPresent() {
-        TableSegment tableSegment1 = createTableSegment("table_1", "tbl_1");
-        TableSegment tableSegment2 = createTableSegment("table_2", "tbl_2");
-        ColumnSegment columnSegment = createColumnSegment();
-        columnSegment.setOwner(new OwnerSegment(0, 10, new IdentifierValue("table_1")));
-        PredicateSegment predicateSegment = createPredicateSegment(columnSegment);
-        Optional<String> actual = new PredicateFinder(mock(RelationMetas.class), Arrays.asList(tableSegment1, tableSegment2)).findTableName(predicateSegment);
-        assertTrue(actual.isPresent());
-        assertThat(actual.get(), is("table_1"));
-    }
-    
-    @Test
-    public void assertFindTableNameWhenColumnSegmentOwnerAbsent() {
-        TableSegment tableSegment1 = createTableSegment("table_1", "tbl_1");
-        TableSegment tableSegment2 = createTableSegment("table_2", "tbl_2");
-        PredicateSegment predicateSegment = createPredicateSegment(createColumnSegment());
-        Optional<String> actual = new PredicateFinder(mock(RelationMetas.class), Arrays.asList(tableSegment1, tableSegment2)).findTableName(predicateSegment);
-        assertFalse(actual.isPresent());
-    }
-    
-    @Test
-    public void assertFindTableNameWhenColumnSegmentOwnerAbsentAndRelationMetasContainsColumn() {
-        TableSegment tableSegment1 = createTableSegment("table_1", "tbl_1");
-        TableSegment tableSegment2 = createTableSegment("table_2", "tbl_2");
-        PredicateSegment predicateSegment = createPredicateSegment(createColumnSegment());
-        RelationMetas relationMetas = mock(RelationMetas.class);
-        when(relationMetas.containsColumn(anyString(), anyString())).thenReturn(true);
-        Optional<String> actual = new PredicateFinder(relationMetas, Arrays.asList(tableSegment1, tableSegment2)).findTableName(predicateSegment);
-        assertTrue(actual.isPresent());
-        assertThat(actual.get(), is("table_1"));
-    }
-    
-    private TableSegment createTableSegment(final String tableName, final String alias) {
-        TableSegment result = new TableSegment(0, 0, new IdentifierValue(tableName));
-        AliasSegment aliasSegment = new AliasSegment(0, 0, new IdentifierValue(alias));
-        result.setAlias(aliasSegment);
-        return result;
-    }
-    
-    private PredicateSegment createPredicateSegment(final ColumnSegment columnSegment) {
-        return new PredicateSegment(0, 0, columnSegment, mock(PredicateRightValue.class));
-    }
-    
-    private ColumnSegment createColumnSegment() {
-        return new ColumnSegment(0, 0, new IdentifierValue("col"));
-    }
-}
